Move BuildingPlacementArgs to CreateBuilding.h
