+++ src/OpenLoco/src/GameCommands/Buildings/RemoveBuilding.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "GameCommands/Buildings/CreateBuilding.h"
 #include "GameCommands/GameCommands.h"
 
 namespace OpenLoco::GameCommands
