+++ src/OpenLoco/src/Windows/TownList.cpp
@@ -1,6 +1,7 @@
 #include "Audio/Audio.h"
 #include "Config.h"
 #include "Drawing/SoftwareDrawingEngine.h"
+#include "GameCommands/Buildings/CreateBuilding.h"
 #include "GameCommands/Buildings/RemoveBuilding.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Town/CreateTown.h"
